requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
networkx>=3.1
numpy>=1.24.0
aiohttp>=3.9.0
asyncio>=3.4.3
redis>=4.5.0
